+++ WebCore/platform/graphics/cairo/FontCairo.cpp
@@ -71,7 +71,7 @@ void Font::drawGlyphs(GraphicsContext* context, const SimpleFontData* font, cons
 
     // Text shadow, inspired by FontMac
     IntSize shadowSize;
-    int shadowBlur = 0;
+    float shadowBlur = 0;
     Color shadowColor;
     bool hasShadow = context->textDrawingMode() == cTextFill &&
         context->getShadow(shadowSize, shadowBlur, shadowColor);
