+++ WebCore/platform/graphics/gtk/FontGtk.cpp
@@ -211,7 +211,7 @@ void Font::drawComplexText(GraphicsContext* context, const TextRun& run, const F
 
     // Text shadow, inspired by FontMac
     IntSize shadowSize;
-    int shadowBlur = 0;
+    float shadowBlur = 0;
     Color shadowColor;
     bool hasShadow = context->textDrawingMode() == cTextFill &&
         context->getShadow(shadowSize, shadowBlur, shadowColor);
